+++ lib/input/reader/websocket_test.go
@@ -95,6 +95,75 @@ func TestWebsocketBasic(t *testing.T) {
 	}
 }
 
+func TestWebsocketOpenMsg(t *testing.T) {
+	expMsgs := []string{
+		"foo",
+		"bar",
+		"baz",
+	}
+
+	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		upgrader := websocket.Upgrader{}
+
+		var ws *websocket.Conn
+		var err error
+		if ws, err = upgrader.Upgrade(w, r, nil); err != nil {
+			return
+		}
+
+		defer ws.Close()
+
+		_, data, err := ws.ReadMessage()
+		if err != nil {
+			t.Fatal(err)
+		}
+		if exp, act := "hello world", string(data); exp != act {
+			t.Errorf("Wrong open message: %v != %v", act, exp)
+		}
+
+		for _, msg := range expMsgs {
+			if err = ws.WriteMessage(websocket.BinaryMessage, []byte(msg)); err != nil {
+				t.Error(err)
+			}
+		}
+	}))
+
+	conf := NewWebsocketConfig()
+	conf.OpenMsg = "hello world"
+	if wsURL, err := url.Parse(server.URL); err != nil {
+		t.Fatal(err)
+	} else {
+		wsURL.Scheme = "ws"
+		conf.URL = wsURL.String()
+	}
+
+	m, err := NewWebsocket(conf, log.New(os.Stdout, log.Config{LogLevel: "NONE"}), metrics.DudType{})
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	if err = m.Connect(); err != nil {
+		t.Fatal(err)
+	}
+
+	for _, exp := range expMsgs {
+		var actMsg types.Message
+		if actMsg, err = m.Read(); err != nil {
+			t.Error(err)
+		} else if act := string(actMsg.Get(0)); act != exp {
+			t.Errorf("Wrong result: %v != %v", act, exp)
+		}
+		if err = m.Acknowledge(nil); err != nil {
+			t.Error(err)
+		}
+	}
+
+	m.CloseAsync()
+	if err = m.WaitForClose(time.Second); err != nil {
+		t.Error(err)
+	}
+}
+
 func TestWebsocketClose(t *testing.T) {
 	closeChan := make(chan struct{})
 	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {